+++ src/main.cpp
@@ -29,7 +29,7 @@
 #include <rest_api_tests/rest_api_tests.h>
 
 #include <libKitsunemimiHanamiCommon/generic_main.h>
-#include <libKitsumiAiSdk/common/websocket_client.h>
+#include <libHanamiAiSdk/common/websocket_client.h>
 
 using Kitsunemimi::Hanami::initMain;
