+++ src/rest_api_tests/kyouko/cluster/cluster_create_test.cpp
@@ -22,7 +22,7 @@
 
 #include "cluster_create_test.h"
 
-#include <libKitsumiAiSdk/cluster.h>
+#include <libHanamiAiSdk/cluster.h>
 
 ClusterCreateTest::ClusterCreateTest(const bool expectSuccess)
           : TestStep(expectSuccess)
