+++ src/rest_api_tests/kyouko/cluster/cluster_delete_test.cpp
@@ -22,7 +22,7 @@
 
 #include "cluster_delete_test.h"
 
-#include <libKitsumiAiSdk/cluster.h>
+#include <libHanamiAiSdk/cluster.h>
 
 ClusterDeleteTest::ClusterDeleteTest(const bool expectSuccess)
           : TestStep(expectSuccess)
