+++ src/rest_api_tests/kyouko/task/image_learn_task_test.cpp
@@ -23,7 +23,7 @@
 #include "image_learn_task_test.h"
 #include <chrono>
 
-#include <libKitsumiAiSdk/task.h>
+#include <libHanamiAiSdk/task.h>
 
 typedef std::chrono::milliseconds chronoMilliSec;
 typedef std::chrono::microseconds chronoMicroSec;
