+++ src/rest_api_tests/kyouko/task/image_request_task_test.cpp
@@ -22,7 +22,7 @@
 
 #include "image_request_task_test.h"
 
-#include <libKitsumiAiSdk/task.h>
+#include <libHanamiAiSdk/task.h>
 
 ImageRequestTaskTest::ImageRequestTaskTest(const bool expectSuccess)
       : TestStep(expectSuccess)
