+++ src/rest_api_tests/kyouko/template/template_delete_test.cpp
@@ -22,7 +22,7 @@
 
 #include "template_delete_test.h"
 
-#include <libKitsumiAiSdk/template.h>
+#include <libHanamiAiSdk/template.h>
 
 TemplateDeleteTest::TemplateDeleteTest(const bool expectSuccess)
             : TestStep(expectSuccess)
