+++ src/rest_api_tests/misaki/project/project_create_test.cpp
@@ -22,7 +22,7 @@
 
 #include "project_create_test.h"
 
-#include <libKitsumiAiSdk/project.h>
+#include <libHanamiAiSdk/project.h>
 
 ProjectCreateTest::ProjectCreateTest(const bool expectSuccess)
     : TestStep(expectSuccess)
