+++ src/rest_api_tests/misaki/user/user_get_test.cpp
@@ -22,7 +22,7 @@
 
 #include "user_get_test.h"
 
-#include <libKitsumiAiSdk/user.h>
+#include <libHanamiAiSdk/user.h>
 
 UserGetTest::UserGetTest(const bool expectSuccess,
                          const std::string &nameOverride)
