+++ src/rest_api_tests/misaki/user/user_list_test.cpp
@@ -22,7 +22,7 @@
 
 #include "user_list_test.h"
 
-#include <libKitsumiAiSdk/user.h>
+#include <libHanamiAiSdk/user.h>
 
 UserListTest::UserListTest(const bool expectSuccess)
     : TestStep(expectSuccess)
