+++ src/rest_api_tests/shiori/datasets/dataset_get_test.cpp
@@ -1,5 +1,5 @@
 /**
- * @file        dataset_create_csv_test.cpp
+ * @file        dataset_get_test.cpp
  *
  * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
  *
@@ -20,32 +20,41 @@
  *      limitations under the License.
  */
 
-#include "dataset_create_csv_test.h"
+#include "dataset_get_test.h"
 
-#include <libKitsunemimiConfig/config_handler.h>
-#include <libKitsumiAiSdk/data_set.h>
+#include <libHanamiAiSdk/data_set.h>
 
-DataSetCreateCsvTest::DataSetCreateCsvTest(const bool expectSuccess)
-          : TestStep(expectSuccess)
+DataSetGetTest::DataSetGetTest(const bool expectSuccess,
+                               const std::string &type,
+                               const std::string &uuidOverride)
+    : TestStep(expectSuccess)
 {
-    m_testName = "create csv data-set";
+    m_testName = "get data-set";
     if(expectSuccess) {
         m_testName += " (success)";
     } else {
         m_testName += " (fail)";
     }
+    m_type = type;
+    m_uuid = uuidOverride;
 }
 
 bool
-DataSetCreateCsvTest::runTest(Kitsunemimi::Json::JsonItem &inputData,
-                              Kitsunemimi::ErrorContainer &error)
+DataSetGetTest::runTest(Kitsunemimi::Json::JsonItem &inputData,
+                        Kitsunemimi::ErrorContainer &error)
 {
-    std::string result;
-    if(Kitsunemimi::Hanami::uploadCsvData(result,
-                                          inputData.get("base_dataset_name").getString(),
-                                          inputData.get("base_inputs").getString(),
-                                          error) != m_expectSuccess)
+    if(m_uuid == "")
     {
+        if(m_type == "learn") {
+            m_uuid = inputData.get("learn_dataset_uuid").getString();
+        } else {
+            m_uuid = inputData.get("request_dataset_uuid").getString();
+        }
+    }
+
+    // get user by name
+    std::string result;
+    if(Kitsunemimi::Hanami::getDataset(result, m_uuid, error) != m_expectSuccess) {
         return false;
     }
 
@@ -59,7 +68,6 @@ DataSetCreateCsvTest::runTest(Kitsunemimi::Json::JsonItem &inputData,
         return false;
     }
 
-    inputData.insert("base_dataset_uuid", jsonItem.get("uuid").getString(), true);
-
     return true;
 }
+
